fix: allow lint-staged to run on empty git repo by disabling backup
Since v10, lint-staged has not been able to run on empty git repos, because the backup stash requires at least the initial commit.

In v10.1.0 an option `--no-stash` was added to make lint-staged skip creating the stash. We can use this same logic to enable running on empty repos, and print a descriptive warning message.