docs: simplify Ignoring Files From .eslintignore. (#1013)
The README advised ESLint >= 7 users to install node-filter-async.
Replace this suggestion with a simpler implementation requiring the same
number of lines of code but no dependencies. Rename the variable
eslintCli to eslint since it is an instance of the ESLint Node.js API as
distinct from the ESLint CLI. Avoid instantiating the API when the files
staged don't necessitate invocation of ESLint.